data_pipeline/requirements.txt
# Api_Milano_Core - Dependencies
# Python >= 3.10

# Data processing
pandas>=2.0.0
geopandas>=0.14.0

# Geometry
shapely>=2.0.0

# HTTP requests
requests>=2.28.0

# CLI enhancements
rich>=13.0.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Data validation
pydantic>=2.5.0

# Type hints
typing-extensions>=4.8.0

# Environment configuration
python-dotenv>=1.0.0
